tests/requirements.txt
# requirements from setup.py
djangocms-attributes-field>=0.1.1
# other requirements
djangocms-helper>=0.9.2,<0.10
tox
coverage
